Move POSTING behavior from TodoForm to App and pass it as a prop to TodoForm. #10
